+++ infrastructure-android/src/main/kotlin/dsl/SourceSets.kt
@@ -1,10 +1,21 @@
 package com.redmadrobot.build.dsl
 
-import com.android.SdkConstants.*
 import com.android.build.api.dsl.AndroidSourceSet
-import com.android.build.api.dsl.CommonExtension
-import org.gradle.api.Incubating
-import org.gradle.kotlin.dsl.get
+import com.redmadrobot.build.android.internal.CommonExtension
+import org.gradle.api.NamedDomainObjectContainer
+
+/** Use `sourceSets.addSharedSourceSetRoot` instead. */
+@Deprecated(
+    "Use sourceSets.addSharedSourceSetRoot(...) instead.",
+    ReplaceWith("this.sourceSets.addSharedSourceSetRoot(variant1, variant2, name)")
+)
+public fun CommonExtension.addSharedSourceSetRoot(
+    variant1: String,
+    variant2: String,
+    name: String = "$variant1${variant2.replaceFirstChar { it.uppercaseChar() }}",
+) {
+    sourceSets.addSharedSourceSetRoot(variant1, variant2, name)
+}
 
 /**
  * Adds source set root with the given [name], shared between [variant1] and [variant2] source sets.
@@ -12,38 +23,34 @@ import org.gradle.kotlin.dsl.get
  * ```
  *  android {
  *      // Here we need to share sources between debug and QA builds
- *      addSharedSourceSetRoot(BUILD_TYPE_DEBUG, BUILD_TYPE_QA)
+ *      sourceSets.addSharedSourceSetRoot(BUILD_TYPE_DEBUG, BUILD_TYPE_QA)
  *
  *      // We can specify name for the source set root if need
- *      addSharedSourceSetRoot(BUILD_TYPE_DEBUG, BUILD_TYPE_QA, name = "debugPanel")
+ *      sourceSets.addSharedSourceSetRoot(BUILD_TYPE_DEBUG, BUILD_TYPE_QA, name = "debugPanel")
  *  }
  * ```
  */
-@Incubating
-public fun CommonExtension<*, *, *, *>.addSharedSourceSetRoot(
+public fun NamedDomainObjectContainer<out AndroidSourceSet>.addSharedSourceSetRoot(
     variant1: String,
     variant2: String,
-    name: String = "$variant1${variant2.capitalize()}",
+    name: String = "$variant1${variant2.replaceFirstChar { it.uppercaseChar() }}",
 ) {
-    val variant1SourceSets = sourceSets[variant1]
-    val variant2SourceSets = sourceSets[variant2]
     val root = "src/$name"
-
-    variant1SourceSets.addRoot(root)
-    variant2SourceSets.addRoot(root)
+    getByName(variant1).addRoot(root)
+    getByName(variant2).addRoot(root)
 }
 
-@Suppress("UnstableApiUsage")
+/** Works similar to [AndroidSourceSet.setRoot], but adds the new [path] to the existing roots. */
 private fun AndroidSourceSet.addRoot(path: String) {
-    java.srcDirs("$path/$FD_JAVA")
-    kotlin.srcDirs("$path/$FD_JAVA", "$path/kotlin")
-    resources.srcDir("$path/$FD_JAVA_RES")
-    res.srcDir("$path/$FD_RES")
-    assets.srcDir("$path/$FD_ASSETS")
-    manifest.srcFile("$path/$FN_ANDROID_MANIFEST_XML")
-    aidl.srcDir("$path/$FD_AIDL")
-    renderscript.srcDir("$path/$FD_RENDERSCRIPT")
+    java.srcDirs("$path/java")
+    kotlin.srcDirs("$path/java", "$path/kotlin")
+    resources.srcDir("$path/resources")
+    res.srcDir("$path/res")
+    assets.srcDir("$path/assets")
+    manifest.srcFile("$path/AndroidManifest.xml")
+    aidl.srcDir("$path/aidl")
+    renderscript.srcDir("$path/rs")
     jniLibs.srcDir("$path/jniLibs")
     shaders.srcDir("$path/shaders")
-    mlModels.srcDir("$path/$FD_ML_MODELS")
+    mlModels.srcDir("$path/ml")
 }