+++ infrastructure-detekt/src/main/kotlin/DetektPlugin.kt
@@ -49,7 +49,7 @@ private fun Project.configureDependencies() {
     }
 
     dependencies {
-        detektPlugins("io.gitlab.arturbosch.detekt:detekt-formatting:1.22.0")
+        detektPlugins("io.gitlab.arturbosch.detekt:detekt-formatting:1.23.6")
     }
 }
 
@@ -135,9 +135,9 @@ private fun Project.configureDetektDiffTask(
         )
 
         val findChangedFiles = tasks.register<CollectGitDiffFilesTask>("findChangedFiles") {
-            projectDir.set(layout.projectDirectory)
-            filterParams.set(changedFilesFilter)
-            branch.set(detektDiffOptions.baseBranch)
+            projectDir = layout.projectDirectory
+            filterParams = changedFilesFilter
+            branch = detektDiffOptions.baseBranch
         }
 
         detektTask(staticAnalyzerSpec, "detektDiff") {
@@ -155,18 +155,18 @@ private inline fun Project.detektTask(
     return tasks.register<Detekt>(name) {
         parallel = true
         config.setFrom(provider { staticAnalyzerSpec.configsDir.get().file("detekt/detekt.yml") })
-        baseline.set(provider { staticAnalyzerSpec.configsDir.getFileIfExists("detekt/baseline.xml") })
+        baseline = provider { staticAnalyzerSpec.configsDir.getFileIfExists("detekt/baseline.xml") }
         setSource(projectDir)
-        reportsDir.set(staticAnalyzerSpec.reportsDir.asFile)
+        reportsDir = staticAnalyzerSpec.reportsDir.asFile
         include("**/*.kt")
         include("**/*.kts")
         exclude("**/res/**")
         exclude("**/build/**")
         exclude("**/.*/**")
         reports {
-            xml.required.set(true)
-            txt.required.set(false)
-            html.required.set(false)
+            xml.required = true
+            txt.required = false
+            html.required = false
         }
         configure()
     }
@@ -178,9 +178,9 @@ private inline fun Project.detektCreateBaselineTask(
     crossinline configure: DetektCreateBaselineTask.() -> Unit,
 ): TaskProvider<DetektCreateBaselineTask> {
     return tasks.register<DetektCreateBaselineTask>(name) {
-        parallel.set(true)
+        parallel = true
         config.setFrom(provider { staticAnalyzerSpec.configsDir.get().file("detekt/detekt.yml") })
-        baseline.set(provider { staticAnalyzerSpec.configsDir.get().file("detekt/baseline.xml") })
+        baseline = provider { staticAnalyzerSpec.configsDir.get().file("detekt/baseline.xml") }
         setSource(projectDir)
         include("**/*.kt")
         include("**/*.kts")
@@ -203,7 +203,7 @@ private inline fun <reified T : SourceTask> Project.extractDetektTaskProviderByT
         val baseExtensions = extensions.getByType<BaseExtension>()
         baseExtensions.checkVariantExists(variantName) { existingVariants ->
             val candidates = existingVariants.joinToString(", ") { variant ->
-                "'${createDetektVariantTaskName(taskSuffix, variant.capitalize(), "All")}'"
+                "'${createDetektVariantTaskName(taskSuffix, variant, "All")}'"
             }
             "Task ${createDetektVariantTaskName(taskSuffix, variantName, "All")} not found in project. " +
                 "Some candidates are: $candidates"
@@ -218,12 +218,12 @@ private inline fun <reified T : SourceTask> Project.extractDetektTaskProviderByT
 }
 
 private fun BaseExtension.checkVariantExists(variantName: String, lazyMessage: (List<String>) -> String) {
-    val requiredVariant = variants?.find { it.name.capitalize() == variantName }
+    val requiredVariant = variants?.find { it.name.capitalized() == variantName }
     checkNotNull(requiredVariant) { lazyMessage.invoke(variants?.map { it.name }.orEmpty()) }
 }
 
 private fun createDetektVariantTaskName(suffix: String, variantName: String, postfix: String = ""): String {
-    return "detekt$suffix$variantName$postfix"
+    return "detekt$suffix${variantName.capitalized()}$postfix"
 }
 
 private fun List<Project>.checkModulesContainDetekt(lazyMessage: (String) -> String) {